chore: move flag to tokenomics
