+++ src/store/modules/AppCommon/reducer/index.ts
@@ -7,7 +7,6 @@ export interface State {
   appInitState: 'loading' | 'success' | 'error';
   appState: AppStateType | null;
   isSplashHidden: boolean;
-  forceStartMining: boolean;
 }
 
 type Actions = ReturnType<
@@ -16,7 +15,6 @@ type Actions = ReturnType<
   | typeof AppCommonActions.APP_INITIALIZED.FAILED.create
   | typeof AppCommonActions.APP_STATE_CHANGE.STATE.create
   | typeof AppCommonActions.UPDATE_SPLASH_VISIBLE_STATE.HIDE.create
-  | typeof AppCommonActions.UPDATE_FORCE_START_MINING.STATE.create
 >;
 
 const INITIAL_STATE: State = {
@@ -24,7 +22,6 @@ const INITIAL_STATE: State = {
   appInitState: 'loading',
   appState: 'active',
   isSplashHidden: false,
-  forceStartMining: false,
 };
 
 export function appCommonReducer(
@@ -47,9 +44,6 @@ export function appCommonReducer(
         break;
       case AppCommonActions.UPDATE_SPLASH_VISIBLE_STATE.HIDE.type:
         draft.isSplashHidden = true;
-        break;
-      case AppCommonActions.UPDATE_FORCE_START_MINING.STATE.type:
-        draft.forceStartMining = action.payload.forceStartMining;
         break;
     }
   });
