+++ src/store/modules/AppCommon/selectors/index.ts
@@ -28,6 +28,3 @@ export const initSuccessSelector = (state: RootState) => {
 };
 export const isSplashHiddenSelector = (state: RootState) =>
   state.appCommon.isSplashHidden;
-
-export const forceStartMiningSelector = (state: RootState) =>
-  state.appCommon.forceStartMining;
