+++ src/store/modules/Tokenomics/sagas/forceMining.ts
@@ -1,10 +1,8 @@
 // SPDX-License-Identifier: ice License 1.0
 
-import {
-  forceStartMiningSelector,
-  isAppActiveSelector,
-} from '@store/modules/AppCommon/selectors';
+import {isAppActiveSelector} from '@store/modules/AppCommon/selectors';
 import {TokenomicsActions} from '@store/modules/Tokenomics/actions';
+import {forceStartMiningSelector} from '@store/modules/Tokenomics/selectors';
 import {put, select} from 'redux-saga/effects';
 
 export function* forceMiningSaga() {
@@ -16,7 +14,6 @@ export function* forceMiningSaga() {
     yield select(forceStartMiningSelector);
 
   if (isAppActive && forceStartMining) {
-    console.log('forceMiningSaga STARTED');
     yield put(TokenomicsActions.START_MINING_SESSION.START.create());
   }
 }
